Modified task 12 to solve the problems of precision loss

+++ 12/9.c
@@ -3,13 +3,13 @@
 
 int main() 
 {
-	float r;
-	float c, s;
-	scanf("%g", &r);
+	double r;
+	double c, s;
+	scanf("%lf", &r);
 	c = 2 * PI * r;
 	s = PI * r *r;
-	printf("%.3f\n", c);
-	printf("%.3f\n", s);
+	printf("%.3lf\n", c);
+	printf("%.3lf\n", s);
 	
 	return 0;
 }
